requirements.txt
praw==7.5.0
pymongo==4.0.1
python-dotenv==0.19.2
schedule==1.1.0